T0

+++ python-server_side_rendering/task_00_intro.py
@@ -1,5 +1,9 @@
 #!usr/bin/python3
 """
+This module is used to generate invitation thanks to a template, using a list
+"""
+import os
+"""
 Definition of generate_invitations function
 """
 
@@ -8,6 +12,7 @@ def generate_invitations(template_content, attendees):
     """
     My base function
     """
+
     if not isinstance(template_content, str):
         raise TypeError("template must be a string!")
     if not isinstance(attendees, list) or\
@@ -28,6 +33,9 @@ def generate_invitations(template_content, attendees):
         ("{event_date}", str(attendee.get("event_date", "N/A")))
         invitation = invitation.replace
         ("{event_location}", str(attendee.get("event_location", "N/A")))
+
+        if os.path.exists(f"output_{i}.txt"):
+            raise FileExistsError("File already exists")
 
         with open(f"output_{i}.txt", 'w', encoding='UTF-8') as file:
             file.write(invitation)
